Do not drop view bounds when converting to bytes.
The `BytesData` really needs to be the underlying view and not the blob. This is not in line with how `Bytes` usually works in js (which possibly should be reconsidered), but it is necessary, because node packs small buffers onto the same `ArrayBuffer`. Also `slice` creates a new `Buffer` backed by the original `ArrayBuffer`.

I recently updated some code to use `hxToBytes`, introducing in weird bugs where the resulting `Bytes` sometimes get padded with junk. I think this is a pretty severe issue.